+++ plugins/physics/src/commonMain/kotlin/com/pandulapeter/kubriko/physics/implementation/physics/dynamics/bodies/PhysicalBody.kt
@@ -2,7 +2,7 @@ package com.pandulapeter.kubriko.physics.implementation.physics.dynamics.bodies
 
 import com.pandulapeter.kubriko.physics.implementation.physics.math.Vec2
 
-class PhysicalBody(x: Double, y: Double) :
+class PhysicalBody(x: Float, y: Float) :
     AbstractPhysicalBody() {
     override var position = Vec2(x, y)
